+++ core/src/main/kotlin/io/bkbn/kompendium/core/plugin/NotarizedRoute.kt
@@ -11,7 +11,6 @@ import io.bkbn.kompendium.core.metadata.PutInfo
 import io.bkbn.kompendium.core.util.Helpers.addToSpec
 import io.bkbn.kompendium.core.util.SpecConfig
 import io.bkbn.kompendium.oas.path.Path
-import io.bkbn.kompendium.oas.path.PathOperation
 import io.bkbn.kompendium.oas.payload.Parameter
 import io.ktor.server.application.ApplicationCallPipeline
 import io.ktor.server.application.Hook
@@ -31,7 +30,6 @@ object NotarizedRoute {
     override var head: HeadInfo? = null
     override var options: OptionsInfo? = null
     override var security: Map<String, List<String>>? = null
-    internal var path: Path? = null
   }
 
   private object InstallHook : Hook<(ApplicationCallPipeline) -> Unit> {
@@ -51,32 +49,22 @@ object NotarizedRoute {
       val spec = application.attributes[KompendiumAttributes.openApiSpec]
       val routePath = route.calculateRoutePath()
       val authMethods = route.collectAuthMethods()
-      pluginConfig.path?.addDefaultAuthMethods(authMethods)
-      require(spec.paths[routePath] == null) {
-        """
-          The specified path ${Parameter.Location.path} has already been documented!
-          Please make sure that all notarized paths are unique
-        """.trimIndent()
-      }
-      spec.paths[routePath] = pluginConfig.path
-        ?: error("This indicates a bug in Kompendium. Please file a GitHub issue!")
+
+      val path = spec.paths[routePath] ?: Path()
+
+      path.parameters = path.parameters?.plus(pluginConfig.parameters) ?: pluginConfig.parameters
+      val serializableReader = application.attributes[KompendiumAttributes.schemaConfigurator]
+
+      pluginConfig.get?.addToSpec(path, spec, pluginConfig, serializableReader, routePath, authMethods)
+      pluginConfig.delete?.addToSpec(path, spec, pluginConfig, serializableReader, routePath, authMethods)
+      pluginConfig.head?.addToSpec(path, spec, pluginConfig, serializableReader, routePath, authMethods)
+      pluginConfig.options?.addToSpec(path, spec, pluginConfig, serializableReader, routePath, authMethods)
+      pluginConfig.post?.addToSpec(path, spec, pluginConfig, serializableReader, routePath, authMethods)
+      pluginConfig.put?.addToSpec(path, spec, pluginConfig, serializableReader, routePath, authMethods)
+      pluginConfig.patch?.addToSpec(path, spec, pluginConfig, serializableReader, routePath, authMethods)
+
+      spec.paths[routePath] = path
     }
-
-    val spec = application.attributes[KompendiumAttributes.openApiSpec]
-    val serializableReader = application.attributes[KompendiumAttributes.schemaConfigurator]
-
-    val path = Path()
-    path.parameters = pluginConfig.parameters
-
-    pluginConfig.get?.addToSpec(path, spec, pluginConfig, serializableReader)
-    pluginConfig.delete?.addToSpec(path, spec, pluginConfig, serializableReader)
-    pluginConfig.head?.addToSpec(path, spec, pluginConfig, serializableReader)
-    pluginConfig.options?.addToSpec(path, spec, pluginConfig, serializableReader)
-    pluginConfig.post?.addToSpec(path, spec, pluginConfig, serializableReader)
-    pluginConfig.put?.addToSpec(path, spec, pluginConfig, serializableReader)
-    pluginConfig.patch?.addToSpec(path, spec, pluginConfig, serializableReader)
-
-    pluginConfig.path = path
   }
 
   private fun Route.calculateRoutePath() = toString().replace(Regex("/\\(.+\\)"), "")
@@ -86,27 +74,4 @@ object NotarizedRoute {
     .map { it.replace("(authenticate ", "").replace(")", "") }
     .map { it.split(", ") }
     .flatten()
-
-  private fun Path.addDefaultAuthMethods(methods: List<String>) {
-    get?.addDefaultAuthMethods(methods)
-    put?.addDefaultAuthMethods(methods)
-    post?.addDefaultAuthMethods(methods)
-    delete?.addDefaultAuthMethods(methods)
-    options?.addDefaultAuthMethods(methods)
-    head?.addDefaultAuthMethods(methods)
-    patch?.addDefaultAuthMethods(methods)
-    trace?.addDefaultAuthMethods(methods)
-  }
-
-  private fun PathOperation.addDefaultAuthMethods(methods: List<String>) {
-    methods.forEach { m ->
-      if (security == null || security?.all { s -> !s.containsKey(m) } == true) {
-        if (security == null) {
-          security = mutableListOf(mapOf(m to emptyList()))
-        } else {
-          security?.add(mapOf(m to emptyList()))
-        }
-      }
-    }
-  }
 }